+++ src/main/kotlin/g2901_3000/s2908_minimum_sum_of_mountain_triplets_i/Solution.kt
@@ -15,7 +15,7 @@ class Solution {
                 for (k in j + 1 until nums.size) {
                     if (nums[i] < nums[j] && nums[k] < nums[j]) {
                         val min = nums[i] + nums[k] + nums[j]
-                        output = min(min.toDouble(), output.toDouble()).toInt()
+                        output = min(min, output)
                     }
                 }
             }
