+++ src/main/kotlin/g2901_3000/s2911_minimum_changes_to_make_k_semi_palindromes/Solution.kt
@@ -28,7 +28,7 @@ class Solution {
         }
         var min = INF
         for (j in (k - 1) * 2 until (i - 1)) {
-            min = min(min.toDouble(), (calc(j, k - 1) + change(j, i)).toDouble()).toInt()
+            min = min(min, calc(j, k - 1) + change(j, i))
         }
         dp[i][k] = min
         return min
