+++ src/main/kotlin/g3301_3400/s3372_maximize_the_number_of_target_nodes_after_connecting_trees_i/Solution.kt
@@ -82,7 +82,7 @@ class Solution {
         run {
             var i = 0
             while (k != 0 && i < m) {
-                max = max(max.toDouble(), b[i][k - 1].toDouble()).toInt()
+                max = max(max, b[i][k - 1])
                 i++
             }
         }
